pyproject.toml
[tool.poetry]
name = "media-organizer"
version = "0.1.0"
description = ""
authors = ["Ed Salisbury <ed.salisbury@gmail.com>"]
readme = "README.md"
packages = [{include = "media_organizer"}]

[tool.poetry.dependencies]
python = "^3.11"
heic-to-jpg = "^0.1.8"
pillow = "^9.3.0"
ffmpeg-python = "^0.2.0"
pillow-heif = "^0.8.0"
python-magic = "^0.4.27"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
